requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
pytz
openai
requests
